@@ -12,92 +12,115 @@
   },
   "homepage": "https://verdaccio.org",
   "main": "build/index.js",
-  "bin": {
-    "verdaccio": "./bin/verdaccio"
-  },
+  "bin": "./bin/verdaccio",
   "funding": {
     "type": "opencollective",
     "url": "https://opencollective.com/verdaccio"
   },
   "dependencies": {
-    "@verdaccio/commons-api": "9.6.1",
-    "@verdaccio/local-storage": "9.7.0",
-    "@verdaccio/readme": "9.7.0",
-    "@verdaccio/streams": "9.6.1",
-    "@verdaccio/ui-theme": "1.11.0",
+    "@verdaccio/commons-api": "9.7.1",
+    "@verdaccio/local-storage": "9.7.4",
+    "@verdaccio/readme": "9.7.3",
+    "@verdaccio/streams": "9.7.2",
+    "@verdaccio/ui-theme": "1.15.1",
     "JSONStream": "1.3.5",
     "async": "3.2.0",
     "body-parser": "1.19.0",
-    "bunyan": "1.8.13",
+    "bunyan": "1.8.15",
     "commander": "3.0.2",
     "compression": "1.7.4",
     "cookies": "0.8.0",
     "cors": "2.8.5",
-    "dayjs": "1.8.28",
-    "envinfo": "7.5.1",
+    "dayjs": "1.10.3",
+    "envinfo": "7.7.3",
     "express": "4.17.1",
     "handlebars": "4.7.6",
-    "http-errors": "1.7.3",
-    "js-yaml": "3.13.1",
+    "http-errors": "1.8.0",
+    "js-yaml": "3.14.1",
     "jsonwebtoken": "8.5.1",
-    "kleur": "4.0.1",
-    "lodash": "4.17.15",
+    "kleur": "4.1.3",
+    "lodash": "4.17.20",
     "lunr-mutable-indexes": "2.3.2",
-    "marked": "0.8.2",
-    "mime": "2.4.6",
+    "marked": "1.2.7",
+    "mime": "2.5.0",
     "minimatch": "3.0.4",
     "mkdirp": "0.5.5",
     "mv": "2.1.1",
     "pkginfo": "0.4.1",
-    "request": "2.87.0",
-    "semver": "6.3.0",
-    "verdaccio-audit": "9.6.1",
-    "verdaccio-htpasswd": "9.6.1"
+    "request": "2.88.0",
+    "semver": "7.3.4",
+    "verdaccio-audit": "9.7.3",
+    "verdaccio-htpasswd": "9.7.2"
   },
   "devDependencies": {
+    "@babel/cli": "7.12.10",
+    "@babel/core": "7.12.10",
+    "@babel/node": "7.12.10",
+    "@babel/plugin-proposal-class-properties": "7.12.1",
+    "@babel/plugin-proposal-decorators": "7.12.12",
+    "@babel/plugin-proposal-export-namespace-from": "7.12.1",
+    "@babel/plugin-proposal-function-sent": "7.12.1",
+    "@babel/plugin-proposal-json-strings": "7.12.1",
+    "@babel/plugin-proposal-numeric-separator": "7.12.7",
+    "@babel/plugin-proposal-object-rest-spread": "7.12.1",
+    "@babel/plugin-proposal-throw-expressions": "7.12.1",
+    "@babel/plugin-syntax-dynamic-import": "7.8.3",
+    "@babel/plugin-syntax-import-meta": "7.10.4",
+    "@babel/plugin-transform-async-to-generator": "7.12.1",
+    "@babel/plugin-transform-classes": "7.12.1",
+    "@babel/plugin-transform-runtime": "7.12.10",
+    "@babel/polyfill": "^7.12.1",
+    "@babel/preset-env": "7.12.11",
+    "@babel/preset-typescript": "7.12.7",
+    "@babel/register": "7.12.10",
+    "@babel/runtime": "7.12.5",
     "@commitlint/cli": "8.3.5",
     "@commitlint/config-conventional": "8.3.4",
-    "@octokit/rest": "16.36.0",
-    "@types/async": "3.2.0",
+    "@octokit/rest": "16.43.2",
+    "@types/async": "3.2.4",
     "@types/bunyan": "1.8.6",
     "@types/express": "4.17.1",
     "@types/http-errors": "1.6.3",
-    "@types/jest": "25.2.1",
-    "@types/lodash": "4.14.149",
+    "@types/jest": "26.0.14",
+    "@types/lodash": "4.14.157",
     "@types/mime": "2.0.1",
     "@types/minimatch": "3.0.3",
     "@types/node": "12.12.21",
     "@types/request": "2.48.3",
     "@types/semver": "6.2.0",
     "@typescript-eslint/eslint-plugin": "2.12.0",
-    "@verdaccio/babel-preset": "^9.6.1",
     "@verdaccio/eslint-config": "^8.5.0",
-    "@verdaccio/types": "^9.7.0",
-    "all-contributors-cli": "6.16.0",
-    "codecov": "3.7.0",
-    "cross-env": "7.0.2",
+    "@verdaccio/types": "^9.7.2",
+    "all-contributors-cli": "6.19.0",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "26.6.3",
+    "babel-loader": "^8.2.2",
+    "babel-plugin-dynamic-import-node": "2.3.3",
+    "babel-plugin-emotion": "10.0.33",
+    "codecov": "3.8.1",
+    "cross-env": "7.0.3",
     "detect-secrets": "1.0.6",
     "eslint": "6.8.0",
-    "fs-extra": "8.1.0",
+    "fs-extra": "9.0.1",
     "get-stdin": "7.0.0",
     "husky": "2.7.0",
     "in-publish": "2.0.1",
-    "jest": "25.3.0",
-    "jest-environment-node": "25.3.0",
+    "jest": "25.5.4",
+    "jest-environment-node": "25.5.0",
     "jest-junit": "9.0.0",
     "lint-staged": "8.2.1",
-    "lockfile-lint": "4.2.2",
-    "nock": "11.7.2",
+    "lockfile-lint": "4.3.7",
+    "nock": "12.0.3",
     "prettier": "1.19.1",
-    "puppeteer": "1.8.0",
+    "puppeteer": "5.5.0",
     "rimraf": "3.0.2",
-    "selfsigned": "1.10.7",
-    "standard-version": "8.0.0",
-    "supertest": "4.0.2",
-    "typescript": "3.7.5",
+    "selfsigned": "1.10.8",
+    "standard-version": "9.1.0",
+    "supertest": "6.1.1",
+    "typescript": "3.9.7",
     "verdaccio": "^4.5.1",
-    "verdaccio-auth-memory": "^9.4.0",
-    "verdaccio-memory": "^9.4.0"
+    "verdaccio-auth-memory": "^9.7.2",
+    "verdaccio-memory": "^9.7.2"
   },
   "keywords": [
     "private",
