@@ -8,7 +8,7 @@ import {CREDENTIALS} from "../functional/config.functional";
 import getPackage from '../functional/fixtures/package';
 
 const buildAuthHeader = (user, pass): string => {
-  return buildToken(TOKEN_BASIC, new Buffer(`${user}:${pass}`).toString('base64'));
+  return buildToken(TOKEN_BASIC, Buffer.from(`${user}:${pass}`).toString('base64'));
 };
 
 export default class Server implements IServerBridge {
