@@ -1,6 +1,5 @@
 import _ from 'lodash';
 import express, { Application } from 'express';
-import bodyParser from 'body-parser';
 import compression from 'compression';
 import cors from 'cors';
 import { HttpError } from 'http-errors';
@@ -28,7 +27,7 @@ const defineAPI = function(config: IConfig, storage: IStorageHandler): any {
   app.use(cors());
 
   // Router setup
-  app.use(log);
+  app.use(log(config));
   app.use(errorReportingMiddleware);
   app.use(function(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer): void {
     res.setHeader('X-Powered-By', config.user_agent);
@@ -46,8 +45,6 @@ const defineAPI = function(config: IConfig, storage: IStorageHandler): any {
   if (config._debug) {
     hookDebug(app, config.self_path);
   }
-
-  app.use(bodyParser.json({ strict: false, limit: config.max_body_size || '10mb' }));
 
   // register middleware plugins
   const plugin_params = {
