@@ -174,96 +174,112 @@ export const LOG_STATUS_MESSAGE = "@{status}, user: @{user}(@{remoteIP}), req: '
 export const LOG_VERDACCIO_ERROR = `${LOG_STATUS_MESSAGE}, error: @{!error}`;
 export const LOG_VERDACCIO_BYTES = `${LOG_STATUS_MESSAGE}, bytes: @{bytes.in}/@{bytes.out}`;
 
-export function log(req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer): void {
-  // logger
-  req.log = logger.child({ sub: 'in' });
+export function log(config: Config) {
+  return function (req: $RequestExtend, res: $ResponseExtend, next: $NextFunctionVer): void {
+    // logger
+    req.log = logger.child({ sub: 'in' });
 
-  const _auth = req.headers.authorization;
-  if (_.isNil(_auth) === false) {
-    req.headers.authorization = '<Classified>';
-  }
+    const _auth = req.headers.authorization;
+    if (_.isNil(_auth) === false) {
+      req.headers.authorization = '<Classified>';
+    }
 
-  const _cookie = req.headers.cookie;
-  if (_.isNil(_cookie) === false) {
-    req.headers.cookie = '<Classified>';
-  }
-
-  req.url = req.originalUrl;
-  req.log.info({ req: req, ip: req.ip }, "@{ip} requested '@{req.method} @{req.url}'");
-  req.originalUrl = req.url;
-
-  if (_.isNil(_auth) === false) {
-    req.headers.authorization = _auth;
-  }
-
-  if (_.isNil(_cookie) === false) {
-    req.headers.cookie = _cookie;
-  }
-
-  let bytesin = 0;
-  req.on('data', function(chunk): void {
-    bytesin += chunk.length;
-  });
-
-  let bytesout = 0;
-  const _write = res.write;
-  // FIXME: res.write should return boolean
-  // @ts-ignore
-  res.write = function(buf): boolean {
-    bytesout += buf.length;
-    /* eslint prefer-rest-params: "off" */
-    // @ts-ignore
-    _write.apply(res, arguments);
-  };
-
-  const log = function(): void {
-    const forwardedFor = req.headers['x-forwarded-for'];
-    const remoteAddress = req.connection.remoteAddress;
-    const remoteIP = forwardedFor ? `${forwardedFor} via ${remoteAddress}` : remoteAddress;
-    let message;
-    if (res._verdaccio_error) {
-      message = LOG_VERDACCIO_ERROR;
-    } else {
-      message = LOG_VERDACCIO_BYTES;
+    const _cookie = req.headers.cookie;
+    if (_.isNil(_cookie) === false) {
+      req.headers.cookie = '<Classified>';
     }
 
     req.url = req.originalUrl;
-    req.log.warn(
-      {
-        request: {
-          method: req.method,
-          url: req.url,
-        },
-        level: 35, // http
-        user: (req.remote_user && req.remote_user.name) || null,
-        remoteIP,
-        status: res.statusCode,
-        error: res._verdaccio_error,
-        bytes: {
-          in: bytesin,
-          out: bytesout,
-        },
-      },
-      message
-    );
-    req.originalUrl = req.url;
-  };
-
-  req.on('close', function(): void {
-    log();
-  });
-
-  const _end = res.end;
-  res.end = function(buf): void {
-    if (buf) {
-      bytesout += buf.length;
+    // avoid log noise data from static content
+    if (req.originalUrl.match(/static/) === null) {
+      req.log.info({req: req, ip: req.ip}, "@{ip} requested '@{req.method} @{req.url}'");
     }
-    /* eslint prefer-rest-params: "off" */
+    req.originalUrl = req.url;
+
+    if (_.isNil(_auth) === false) {
+      req.headers.authorization = _auth;
+    }
+
+    if (_.isNil(_cookie) === false) {
+      req.headers.cookie = _cookie;
+    }
+
+    let bytesin = 0;
+    if (config?.experiments?.bytesin_off !== true) {
+      req.on('data', function(chunk): void {
+        bytesin += chunk.length;
+      });
+    }
+
+    let bytesout = 0;
+    const _write = res.write;
+    // FIXME: res.write should return boolean
     // @ts-ignore
-    _end.apply(res, arguments);
-    log();
-  };
-  next();
+    res.write = function(buf): boolean {
+      bytesout += buf.length;
+      /* eslint prefer-rest-params: "off" */
+      // @ts-ignore
+      _write.apply(res, arguments);
+    };
+
+    let logHasBeenCalled = false;
+    const log = function(): void {
+      if (logHasBeenCalled) {
+        return;
+      }
+      logHasBeenCalled = true;
+
+      const forwardedFor = req.headers['x-forwarded-for'];
+      const remoteAddress = req.connection.remoteAddress;
+      const remoteIP = forwardedFor ? `${forwardedFor} via ${remoteAddress}` : remoteAddress;
+      let message;
+      if (res._verdaccio_error) {
+        message = LOG_VERDACCIO_ERROR;
+      } else {
+        message = LOG_VERDACCIO_BYTES;
+      }
+
+      req.url = req.originalUrl;
+      // avoid log noise data from static content
+      if (req.url.match(/static/) === null) {
+        req.log.warn(
+            {
+              request: {
+                method: req.method,
+                url: req.url,
+              },
+              level: 35, // http
+              user: (req.remote_user && req.remote_user.name) || null,
+              remoteIP,
+              status: res.statusCode,
+              error: res._verdaccio_error,
+              bytes: {
+                in: bytesin,
+                out: bytesout,
+              },
+            },
+            message
+        );
+        req.originalUrl = req.url;
+      }
+    }
+
+    req.on('close', function(): void {
+      log();
+    });
+
+    const _end = res.end;
+    res.end = function(buf): void {
+      if (buf) {
+        bytesout += buf.length;
+      }
+      /* eslint prefer-rest-params: "off" */
+      // @ts-ignore
+      _end.apply(res, arguments);
+      log();
+    };
+    next();
+  }
 }
 
 // Middleware
@@ -272,7 +288,7 @@ export function errorReportingMiddleware(req: $RequestExtend, res: $ResponseExte
     res.report_error ||
     function(err: VerdaccioError): void {
       if (err.status && err.status >= HTTP_STATUS.BAD_REQUEST && err.status < 600) {
-        if (_.isNil(res.headersSent) === false) {
+        if (!res.headersSent) {
           res.status(err.status);
           next({ error: err.message || API_ERROR.UNKNOWN_ERROR });
         }
