@@ -3,7 +3,6 @@ import {runVerdaccio} from '../../utils/process';
 import {installVerdaccio} from "../__partials/npm_commands";
 
 describe('verdaccio info', ()=> {
-	jest.setTimeout(90000);
 	// @ts-ignore
 	const tempRootFolder = global.__namespace.getItem('dir-root');
 	const verdaccioInstall = path.join(tempRootFolder, 'verdaccio-root-info');
@@ -11,8 +10,7 @@ describe('verdaccio info', ()=> {
 
 	beforeAll(async () => {
 		await installVerdaccio(verdaccioInstall);
-	});
-
+	}, 30000);
 
 	test('should run verdaccio info command', async () => {
 		const pathVerdaccioModule = require.resolve('verdaccio/bin/verdaccio', {
@@ -21,7 +19,7 @@ describe('verdaccio info', ()=> {
 		const hasMatch = await runVerdaccio(pathVerdaccioModule, verdaccioInstall, ['--info'], /Environment/);
 
 		expect(hasMatch.ok).toBeTruthy();
-	});
+	}, 20000);
 
 	afterAll(() => {
 		registryProcess.kill();
