@@ -2,6 +2,8 @@ import request from 'supertest';
 import _ from 'lodash';
 import path from 'path';
 import rimraf from 'rimraf';
+import nock from 'nock';
+import { Readable } from 'stream';
 
 import configDefault from '../../partials/config';
 import publishMetadata from '../../partials/publish-api';
@@ -32,6 +34,12 @@ import {
   generateVersion,
 } from '../../__helper/utils';
 
+const sleep = (delay) => {
+  return new Promise(resolve => {
+    setTimeout(resolve, delay)
+  });
+}
+
 require('../../../../src/lib/logger').setup([
   { type: 'stdout', format: 'pretty', level: 'warn' }
 ]);
@@ -51,11 +59,11 @@ const putVersion = (app, name, publishMetadata) => {
 
 describe('endpoint unit test', () => {
   let app;
+  const mockServerPort = 55549;
   let mockRegistry;
 
   beforeAll(function(done) {
     const store = path.join(__dirname, '../../partials/store/test-storage-api-spec');
-    const mockServerPort = 55549;
     rimraf(store, async () => {
       const configForTest = configDefault({
         auth: {
@@ -74,6 +82,12 @@ describe('endpoint unit test', () => {
         uplinks: {
           npmjs: {
             url: `http://${DOMAIN_SERVERS}:${mockServerPort}`
+          },
+          socketTimeout: {
+            url: `http://some.registry.timeout.com`,
+            max_fails: 2,
+            timeout: '1s',
+            fail_timeout: '1s'
           }
         },
         logs: [
@@ -91,6 +105,10 @@ describe('endpoint unit test', () => {
     mockRegistry[0].stop();
     done();
   });
+
+  afterEach(() => {
+    nock.cleanAll();
+  })
 
   describe('Registry API Endpoints', () => {
 
@@ -354,6 +372,39 @@ describe('endpoint unit test', () => {
             done();
           });
       });
+
+      test('should fails with socket time out fetch tarball timeout package from remote uplink', async () => {
+        const timeOutPkg = generatePackageMetadata('timeout', '1.5.1');
+        const responseText = 'fooooooooooooooooo';
+        const readable = Readable.from([responseText]);
+        timeOutPkg.versions['1.5.1'].dist.tarball = 'http://some.registry.timeout.com/timeout/-/timeout-1.5.1.tgz';
+        nock('http://some.registry.timeout.com')
+          .get('/timeout')
+          .reply(200, timeOutPkg);
+        nock('http://some.registry.timeout.com')
+          .get('/timeout/-/timeout-1.5.1.tgz')
+          .twice()
+          .socketDelay(50000)
+        .reply(200);
+        nock('http://some.registry.timeout.com')
+        .get('/timeout/-/timeout-1.5.1.tgz')
+        .reply(200, () => readable);
+        const agent = request.agent(app);
+        await agent
+          .get('/timeout/-/timeout-1.5.1.tgz')
+          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.OCTET_STREAM)
+          .expect(HTTP_STATUS.INTERNAL_ERROR);
+        await agent
+          .get('/timeout/-/timeout-1.5.1.tgz')
+          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.OCTET_STREAM)
+          .expect(HTTP_STATUS.INTERNAL_ERROR);
+        await sleep(2000);
+        // await agent
+        await agent
+          .get('/timeout/-/timeout-1.5.1.tgz')
+          .expect(HEADER_TYPE.CONTENT_TYPE, HEADERS.OCTET_STREAM)
+          .expect(HTTP_STATUS.OK);
+      }, 10000);
 
       test('should fetch jquery specific version package from remote uplink', (done) => {
 
